long port => int port, as the c source uses! (Miklos Nemeth found this)

+++ lib/connect.h
@@ -28,7 +28,7 @@ int Curl_nonblock(int socket,    /* operate on this */
 
 CURLcode Curl_connecthost(struct connectdata *conn,
                           Curl_addrinfo *host, /* connect to this */
-                          long port,      /* connect to this port number */
+                          int port,       /* connect to this port number */
                           int *sockconn,  /* not set if error is returned */
                           Curl_ipconnect **addr /* the one we used */
                           ); /*  index we used */
